Configure client application
Create project with [create-nuxt-app][1] using [yarn][2].

Create [Dockerfile][3] and [docker-compose.yml][4] to run a container
with the web server and allow connecting it to the DB and API. Create
file .env for further configuration of the service.

Scaffold an HTML [holy grail][5] using CSS [Grid][6] and [Flexbox][7]
to create the structure for the website.

Use [Normalize][8] to standardize style between browsers.
Include [SCSS][9] to style components and create file custom.scss
to define variables with project's colors.

Include the JS library [Axios][10] via Nuxt's modules to make
the communication with the server.

Enable [Vuex][11] via [Nuxt's integration][12] and create a
[Nuxt middlewares][13] to handle authentication and authorization.

In order to start the container, run `docker-compose up` from the
directory 'client'. Nuxt will serve the website at port 5000.

Signed-off-by: Renato Cordeiro Ferreira <renato.cferreira@hotmail.com>

[1]: https://github.com/nuxt-community/create-nuxt-app
[2]: https://yarnpkg.com
[3]: https://docs.docker.com/engine/reference/builder/
[4]: https://docs.docker.com/compose/
[5]: https://en.wikipedia.org/wiki/Holy_grail_(web_design)
[6]: https://css-tricks.com/snippets/css/complete-guide-grid/
[7]: https://css-tricks.com/snippets/css/a-guide-to-flexbox/
[8]: https://necolas.github.io/normalize.css/
[9]: https://sass-lang.com/
[10]: https://axios.nuxtjs.org/
[11]: https://vuex.vuejs.org/
[12]: https://nuxtjs.org/guide/vuex-store
[13]: https://nuxtjs.org/guide/routing#middleware